Made an alias for ticketing.TicketingHandlar as more.DatabaseHandler. Move CouchDBBackend in to more.py
